+++ source/digioz.Forum/digioz.Forum/Models/ForumWarning.cs
@@ -9,7 +9,7 @@ public partial class ForumWarning
 
     public int UserId { get; set; }
 
-    public int PostId { get; set; }
+    public long PostId { get; set; }
 
     public int LogId { get; set; }
